Fixed vertical cloud bug
...also fixed physics bug in this commit, cuz why not

+++ src/graphics/background.cpp
@@ -45,8 +45,7 @@ vec2 Background::pos2() {
 }
 
 void Background::set_pos1(const vec2 &pos) {
-    sp1_.set_pos((int)pos.x, (int)pos.y);
-//    sp1_.set_pos(pos.x, pos.y);
+    sp1_.set_pos(pos.x, pos.y);
 }
 
 void Background::set_pos1(float x, float y) {
@@ -54,8 +53,7 @@ void Background::set_pos1(float x, float y) {
 }
 
 void Background::set_pos2(const vec2 &pos) {
-    sp2_.set_pos((int)pos.x, (int)pos.y);
-//    sp2_.set_pos(pos.x, pos.y);
+    sp2_.set_pos(pos.x, pos.y);
 }
 
 void Background::set_pos2(float x, float y) {
